Support non-fatal errors
Prior to this, if there is an evaluation error, it immediately stops the rest of the evaluation. In other words, all errors are "fatal" errors.

Support non-fatal errors by supporting generation and display of errors that do not stop the rest of the evaluation.

This change makes reusing an existing target name a non-fatal error.

Fixes #61.