Add root-file setting.
Add a new optional `Root File` setting to control the FASTBuild root `fbuild.bff` file. The root file normally does not need to be specified in the settings because it is detected automatically when it is in a parent directory. But it is necessary to set when invoking FASTBuild with the `-config <path>` command line option, which overrides the root file.

For example, if you have the following files and run FASTBuild with `fbuild -config <workspace_root>/Projects/fbuild.bff`:
```
<workspace_root>/Projects/fbuild.bff
<workspace_root>/Projects/HelloWorld/HelloWorld.bff
<workspace_root>/External/MSVC/MSVC.bff
```
In this example, `MSVC.bff` by default cannot find the root `fbuild.bff`. So it is necessary to use the `Root File` setting to manually specify the root.

This new setting is the first that the user can set to an invalid value (an invalid path). Handle that by adding settings validation and output of settings-error diagnostics.

This is also the first time that changing a setting can invalidate already-evaluated data. So clear the cached results when the settings change.

Now that getting the root file relies on settings, which arrives async, various other functions need to also become async.

Fixes #82.